# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CineHall API",
        "description": "API documentation for CineHall Cinema Management System",
        "contact": {
            "email": "support@cinehall.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "CineHall API Server V1"
        }
    ],
    "paths": {
...
        "/sessions": {
            "get": {
                "tags": [
                    "Sessions"
                ],
                "summary": "List all film sessions",
                "operationId": "2be113935b93a4d92ee177f9e82e8e1b",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Sessions"
                ],
                "summary": "Create a new film session",
                "operationId": "e0eb8f6fe332ba08ace4b1401932c406",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "film_id",
                                    "room_id",
                                    "language",
                                    "start_time",
                                    "end_time",
                                    "price"
                                ],
                                "properties": {
                                    "film_id": {
                                        "type": "integer"
                                    },
                                    "room_id": {
                                        "type": "integer"
                                    },
                                    "language": {
                                        "type": "string"
                                    },
                                    "start_time": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "end_time": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Session created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Session"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/sessions/{film_session}": {
            "get": {
                "tags": [
                    "Sessions"
                ],
                "summary": "Show a specific film session",
                "operationId": "d8ef4f006ddee4a227ec8275efd5f986",
                "parameters": [
                    {
                        "name": "film_session",
                        "in": "path",
                        "description": "Session ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Session"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Session not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Sessions"
                ],
                "summary": "Update an existing film session",
                "operationId": "19363b065c76b3b258851fca90747626",
                "parameters": [
                    {
                        "name": "film_session",
                        "in": "path",
                        "description": "Session ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "film_id": {
                                        "type": "integer"
                                    },
                                    "room_id": {
                                        "type": "integer"
                                    },
                                    "language": {
                                        "type": "string"
                                    },
                                    "start_time": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "end_time": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Session updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Session"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Sessions"
                ],
                "summary": "Delete a film session",
                "operationId": "52faf3022b7e70313e7eddfe757f029d",
                "parameters": [
                    {
                        "name": "film_session",
                        "in": "path",
                        "description": "Session ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Session deleted successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/transactions/stripe": {
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Create a Stripe Checkout Session",
                "operationId": "92ab097055f7bcb2974714bba82b3082",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "reservation_id",
                                    "amount"
                                ],
                                "properties": {
                                    "reservation_id": {
                                        "type": "integer"
                                    },
                                    "amount": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Stripe session created"
                    },
                    "400": {
                        "description": "Invalid request or already paid"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/transactions/stripe/success": {
            "get": {
                "tags": [
                    "Payments"
                ],
                "summary": "Handle successful Stripe payment",
                "operationId": "e6bf364c8f9cf1456f6bef536876438b",
                "parameters": [
                    {
                        "name": "reservation_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "session_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment confirmed"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Session": {
                "title": "Session Model",
                "description": "Session (Film Showing) model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true,
                        "example": 1
                    },
                    "film_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "room_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "language": {
                        "type": "string",
                        "example": "English"
                    },
                    "start_time": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-03-22T18:00:00Z"
                    },
                    "end_time": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-03-22T20:00:00Z"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 12.5
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter JWT token to access protected endpoints",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}